Added new features to 9DoF
Added a zero calibration system that calibrates X and Y of accelerometer
and gyro, and Z of gyro to zero on startup and on command. Added a data
transmission mode as well that can select what data it sends. Moved some
of the code in the main file to header files to make it easier to read.